Remove Event from types again
