+++ sdk/language-support/ts/daml-ledger/index.ts
@@ -1,13 +1,10 @@
 // Copyright (c) 2025 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 import {
-  ArchiveEvent,
   Choice,
   ChoiceFrom,
   DisclosedContract,
   ContractId,
-  CreateEvent,
-  Event,
   List,
   Party,
   Template,
@@ -19,6 +16,20 @@ import fetch from "cross-fetch";
 import { EventEmitter } from "events";
 import WebSocket from "isomorphic-ws";
 import _, { isUndefined } from "lodash";
+
+/**
+ * The result of a ``query`` against the ledger.
+ *
+ * @typeparam T The contract template type of the query.
+ * @typeparam K The contract key type of the query.
+ * @typeparam I The template id type.
+ */
+export type QueryResult<T extends object, K, I extends string> = {
+  /** Contracts matching the query. */
+  contracts: readonly CreateEvent<T, K, I>[];
+  /** Indicator for whether the query is executing. */
+  loading: boolean;
+};
 
 /**
  * Full information about a Party.
@@ -93,6 +104,50 @@ export type PackageId = string;
 const decode = <R>(decoder: jtv.Decoder<R>, data: unknown): R => {
   return jtv.Result.withException(decoder.run(data));
 };
+
+/**
+ * A newly created contract.
+ *
+ * @typeparam T The contract payload type.
+ * @typeparam K The contract key type.
+ * @typeparam I The contract type id.
+ *
+ */
+export type CreateEvent<
+  T extends object,
+  K = unknown,
+  I extends string = string,
+> = {
+  templateId: I;
+  contractId: ContractId<T>;
+  signatories: List<Party>;
+  observers: List<Party>;
+  key: K;
+  payload: T;
+};
+
+/**
+ * An archived contract.
+ *
+ * @typeparam T The contract template or interface type.
+ * @typeparam I The template or interface id.
+ */
+export type ArchiveEvent<T extends object, I extends string = string> = {
+  templateId: I;
+  contractId: ContractId<T>;
+};
+
+/**
+ * An event is either the creation or archival of a contract.
+ *
+ * @typeparam T The contract template type.
+ * @typeparam K The contract key type.
+ * @typeparam I The contract id type.
+ */
+export type Event<T extends object, K = unknown, I extends string = string> =
+  | { created: CreateEvent<T, K, I>; matchedQueries: number[] }
+  | { created: CreateEvent<T, K, I> }
+  | { archived: ArchiveEvent<T, I> };
 
 function lookupTemplateOrUnknownInterface(
   templateId: string,
